+++ src/providers/commons.ts
@@ -1,0 +1,24 @@
+import { Injectable } from '@angular/core';
+import { AlertController } from 'ionic-angular';
+
+@Injectable()
+export class CuentasCobrarProvider {
+
+	boxesRef: AngularFireList<any>;
+	boxesStatusRef: AngularFireList<any>;
+	typeBoxRef: AngularFireList<any>;
+	typePayRef: AngularFireList<any>;
+	tenantsRef: AngularFireList<any>;
+
+	constructor(public alertCtrl: AlertController) {
+	}
+
+  public createAlert(title, message) {
+    let alert = this.alertCtrl.create({
+      title: title,
+      subTitle: message,
+      buttons: ['Dismiss']
+    });
+    alert.present();
+  }
+}
